Testing

+++ Dev/app/src/main/java/com/example/gruppe30in2000/FavCity/SwipeController.kt
@@ -2,6 +2,7 @@ package com.example.gruppe30in2000.FavCity
 
 import android.annotation.SuppressLint
 import android.graphics.Canvas
+import android.provider.MediaStore.Audio.Playlists.Members.moveItem
 import android.support.v7.app.AlertDialog
 import android.support.v7.widget.RecyclerView
 import android.support.v7.widget.helper.ItemTouchHelper
@@ -15,6 +16,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.widget.Button
 import com.example.gruppe30in2000.R
+import kotlinx.android.synthetic.main.city_element.view.*
 
 
 enum class ButtonsState {
@@ -37,29 +39,13 @@ open class SwipeController : Callback() {
         viewHolder: RecyclerView.ViewHolder,
         target: RecyclerView.ViewHolder
     ): Boolean {
-//                moveItem(viewHolder.adapterPosition, target.adapterPosition)
+        Log.e("Swipeconntroller", "Moving")
+        moveItem(viewHolder.adapterPosition, target.adapterPosition)
         return true
     }
 
     override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
-//                Log.e("onSwiped:", "Swiping")
-//                val dialogBuilder = AlertDialog.Builder(recyclerView.context) // make a dialog builder
-//                val dialogView = LayoutInflater.from(recyclerView.context).inflate(R.layout.delete_alert, null) // get the dialog xml view
-//                dialogBuilder.setView(dialogView) // set the view into the builder
-//                val alertDialog = dialogBuilder.create()
-//                alertDialog.show()
-//
-//                val okButton = dialogView.findViewById<Button>(R.id.ok_button)
-//                val cancelButton = dialogView.findViewById<Button>(R.id.cancel_button)
-//
-//                okButton.setOnClickListener {
-//                    deleteItem(viewHolder.adapterPosition)
-//                    alertDialog.hide()
-//                }
-//
-//                cancelButton.setOnClickListener {
-//                    alertDialog.hide()
-//                }
+
     }
 
     override fun convertToAbsoluteDirection(flags: Int, layoutDirection: Int): Int {
@@ -69,6 +55,7 @@ open class SwipeController : Callback() {
         }
         return super.convertToAbsoluteDirection(flags, layoutDirection)
     }
+
     override fun onChildDraw(
         c: Canvas,
         recyclerView: RecyclerView,
@@ -91,34 +78,30 @@ open class SwipeController : Callback() {
         dX: Float, dY: Float,
         actionState: Int, isCurrentlyActive: Boolean
     ) {
-
         recyclerView.setOnTouchListener(object : View.OnTouchListener {
             override fun onTouch(v: View, event: MotionEvent): Boolean {
                 swipeBack = event.action == MotionEvent.ACTION_CANCEL || event.action == MotionEvent.ACTION_UP
                 if (swipeBack) {
                     // If swipe left, display dialog to delete
                     if (dX <= buttonWidth) {
-                        Log.e("onSwiped:", "Swiping")
-                        val dialogBuilder = AlertDialog.Builder(recyclerView.context) // make a dialog builder
-                        val dialogView = LayoutInflater.from(recyclerView.context).inflate(R.layout.delete_alert, null) // get the dialog xml view
-                        dialogBuilder.setView(dialogView) // set the view into the builder
-                        val alertDialog = dialogBuilder.create()
-                        alertDialog.show()
+                        if (isCurrentlyActive) {
+                            val dialogBuilder = AlertDialog.Builder(recyclerView.context) // make a dialog builder
+                            val dialogView = LayoutInflater.from(recyclerView.context).inflate(R.layout.delete_alert, null) // get the dialog xml view
+                            dialogBuilder.setView(dialogView) // set the view into the builder
+                            val alertDialog = dialogBuilder.create()
+                            alertDialog.show()
 
-                        val okButton = dialogView.findViewById<Button>(R.id.ok_button)
-                        val cancelButton = dialogView.findViewById<Button>(R.id.cancel_button)
+                            val okButton = dialogView.findViewById<Button>(R.id.ok_button)
+                            val cancelButton = dialogView.findViewById<Button>(R.id.cancel_button)
 
-                        Log.e("adapterPosition", viewHolder.adapterPosition.toString())
-//                        if (viewHolder.adapterPosition > 0) {
-//                        }
+                            okButton.setOnClickListener {
+                                deleteItem(viewHolder.adapterPosition)
+                                alertDialog.hide()
+                            }
 
-                        okButton.setOnClickListener {
-                            deleteItem(viewHolder.adapterPosition)
-                            alertDialog.hide()
-                        }
-
-                        cancelButton.setOnClickListener {
-                            alertDialog.hide()
+                            cancelButton.setOnClickListener {
+                                alertDialog.hide()
+                            }
                         }
                     }
                 }
@@ -127,6 +110,9 @@ open class SwipeController : Callback() {
         })
     }
     open fun deleteItem(pos: Int) {
+        return
+    }
+    open fun moveItem(oldPos: Int, newPos: Int) {
         return
     }
 }
